requirements.txt
torch>=2.4.0
torchvision
diffusers>=0.32.0
transformers>=4.45.0
accelerate>=0.34.0
safetensors
sentencepiece
protobuf
ftfy
imageio
imageio-ffmpeg
opencv-python-headless
numpy
Pillow
PyQt6>=6.6.0
psutil
